Created script and module for writing INPUT_RRTM for RRTMG-LW
The script is to be the place where the user enters all
values for input parameters for RRTMG-LW.  The script is adapted directely
from AER's instructions, so each parameter has a description next to it.

After the user has specified all values, 'python rrtmg_lw.py' writes out
INPUT_RRTM, which is the input file (or tape) used by RRTMG-LW's executable.

The module rtmtools.rrtmg.lw.input contains routines for writing
out the input tape to the required format.

	modified:   rtmtools/rtmtools/rrtmg/lw/input.py
	modified:   rtmtools/rtmtools/rrtmg/scripts/rrtmg_lw.py