lblrtm_sw.py: make (wavenumber, pressure, attributes) Panel for output
The problem here is that 'wavenumber' varies with 'pressure'.  Different
observer pressure yields a different set of wavenumbers.  This means that results
from different observer pressure are concatentated together, there are a lot
of NaNs, for wavenumbers at which a particular pressure does not have the flux or heating rate value.